+++ sdk/resourcemanager/azure-resourcemanager-resources/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.0.0-beta.5 (Unreleased)
 
+- Added `AggregatedManagementException` exception class for aggregation of multiple `ManagementException`.
 
 ## 2.0.0-beta.4 (2020-09-02)